requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking
#
alabaster==0.7.12
    # via sphinx
asgiref==3.6.0
    # via django
autopep8==2.0.1
    # via django-silk
babel==2.11.0
    # via sphinx
certifi==2022.12.7
    # via
    #   pipenv
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via requests
cryptography==39.0.0
    # via jwcrypto
deprecated==1.2.13
    # via jwcrypto
distlib==0.3.6
    # via virtualenv
dj-database-url==1.2.0
    # via -r requirements.in
django==4.1.5
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-appconf
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-oauth-toolkit
    #   django-request-logging
    #   django-silk
    #   djangorestframework
django-appconf==1.0.5
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-oauth-toolkit==2.2.0
    # via -r requirements.in
django-request-logging==0.7.5
    # via -r requirements.in
django-silk==5.0.2
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
djangorestframework-jwt==1.11.0
    # via -r requirements.in
docutils==0.19
    # via sphinx
filelock==3.9.0
    # via virtualenv
fortune==1.1.0
    # via -r requirements.in
gprof2dot==2022.7.29
    # via django-silk
grizzled-python==2.2.0
    # via fortune
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
imagesize==1.4.1
    # via sphinx
jinja2==3.1.2
    # via
    #   django-silk
    #   sphinx
jwcrypto==1.4.2
    # via django-oauth-toolkit
markupsafe==2.1.1
    # via jinja2
oauthlib==3.2.2
    # via django-oauth-toolkit
packaging==23.0
    # via sphinx
pipenv==2022.12.19
    # via -r requirements.in
platformdirs==2.6.2
    # via virtualenv
psycopg2-binary==2.9.5
    # via -r requirements.in
pycodestyle==2.10.0
    # via autopep8
pycparser==2.21
    # via cffi
pydotplus==2.0.2
    # via -r requirements.in
pygments==2.14.0
    # via sphinx
pyjwt==1.7.1
    # via djangorestframework-jwt
pyparsing==3.0.9
    # via pydotplus
python-dateutil==2.8.2
    # via django-silk
pytz==2022.7
    # via
    #   babel
    #   djangorestframework
requests==2.28.2
    # via
    #   django-oauth-toolkit
    #   django-silk
    #   sphinx
six==1.16.0
    # via python-dateutil
snowballstemmer==2.2.0
    # via sphinx
sphinx==6.1.3
    # via -r requirements.in
sphinxcontrib-applehelp==1.0.3
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.3
    # via
    #   django
    #   django-silk
urllib3==1.26.14
    # via requests
virtualenv==20.17.1
    # via pipenv
virtualenv-clone==0.5.7
    # via pipenv
whitenoise==6.3.0
    # via -r requirements.in
wrapt==1.14.1
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
